requirements.txt
python-telegram-bot==13.15
urllib3==1.26.18
six==1.16.0
requests